docker-compose.yml: restart unless-stopped
Adding this item to the nginx, wiki, and wiki-db services
accomplishes two things:

1. It ensures that all essential wiki services are online unless
   we have explicitly stopped them, which reduces the likelihood
   of service disruption.

2. It solves (in an unsatisfying way) the problem of nginx failing
   to start unless the wiki service is already started.

For 2, the issue is that there is a degeneracy between logical
dependencies and functional dependencies in docker-compose. The wiki
service functionally depends on both the nginx and wiki-db services
(these need to be running for the wiki to be functional). However,
the nginx service logically depends on wiki (since the reference
to the wiki service in nginx.conf causes an error unless this
service is already exposed on the local network).

However, specifying both the logical and functional dependencies
would result in a "circular dependencies" error from docker-compose.

A workaround is to abuse the "restart: unless-stopped" condition
so that, even though nginx will fail on its first try to start,
it will succeed a little later as the wiki service spins up
afterwards. This allows us to specify on the functional dependencies.
It is unsatisfying because the nginx container will report errors
until it successfully starts up.

The alternative is to specify the logical dependencies instead, but
then you must remember to specify all services on the command-line
in order for the services to be functional, i.e.
`docker-compose up --build -d wiki nginx`.

See https://github.com/docker/compose/issues/2589 for a relevant
feature request.